Improve SpecParser interface for CLI use
