+++ request-manager/spec/parser_test.go
@@ -48,3 +48,18 @@ func TestWarnParseSpec(t *testing.T) {
 		}
 	}
 }
+
+/* ========================================================================= */
+// Make sure checks run correctly
+
+func TestRunChecks(t *testing.T) {
+	sequencesFile := specsDir + "decomm.yaml"
+	allSpecs, _, _ := ParseSpec(sequencesFile)
+	err, warn := RunChecks(allSpecs)
+	if len(err) > 0 {
+		t.Errorf("decomm.yaml failed check, expected success: %v", err)
+	}
+	if len(warn) > 0 {
+		t.Errorf("decomm.yaml produced warnings, expected none: %v", warn)
+	}
+}
